Extract out `ProgressPrinter`
This is a preparation before supporting 'clang-cl' for Windows (#1179).

Our plan is to download LLVM for Windows in 'update_deps.py', and in
that process we would like to use 'ProgressPrinter' to show progress for
both downloading the archive and extracting it.

To avoid code duplicate between 'build_qt.py' and 'update_deps.py',
let's extract 'ProgressPrinter' into a new file first.

This is a mechanical refactoring commit. There must be no observable
behavior change.
